# backend/doc/swagger.yaml
swagger: '2.0'
info:
  title: Appli Atelier
  description: Backend de l'appli Atelier de Maintenance. Toutes les données sont encapsulées dans une réponse conforme au standard "JSend" (https://labs.omniti.com/labs/jsend)
  version: 0.0.1
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users/login:
    post:
      summary: Login user
      description: This WS is used to login users
      parameters:
        - in: body
          name: credentials
          description: The user credentials
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                description: Nom d'utilisateur
                example: ocp
              password:
                type: string
                description: Mot de passe
                example: password
      responses:
        '200':
          description: Login réussi
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                description: statut selon le standard Jsend
                enum: [
                  "success",
                  "fail",
                  "error"
                ]
              code:
                type: integer
                description: code numérique en cas d'erreur selon le standard Jsend
              message:
                type: string
                description: message en cas d'erreur selon le standard Jsend
              data:
                $ref: '#/definitions/User'
        '401':
          description: Login échoué
  /users:
    get:
      summary: Liste les utilisateurs
      description: >-
        Renvoie la liste des utilisateurs, éventuellement filtrée selon
        l'utilisateur connecté et les paramètres de filtre
      parameters:
        - in: query
          name: section
          type: integer
          description: filtre par Section
        - in: query
          name: role
          type: integer
          description: filtre par Rôle
        - in: query
          name: search
          type: string
          description: 'chaîne à rechercher dans le Username, Nom ou Prénom des utilisateurs'
      responses:
        '200':
          description: Liste d'utilisateurs correspondant
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                description: statut selon le standard Jsend
                enum: [
                  "success",
                  "fail",
                  "error"
                ]
              code:
                type: integer
                description: code numérique en cas d'erreur selon le standard Jsend
              message:
                type: string
                description: message en cas d'erreur selon le standard Jsend
              data:
                type: array
                items:
                  $ref: '#/definitions/User'
        '401':
          description: Accès non autorisé
...
definitions:
  Role:
    type: object
    description: Role d'un utilisateur
    properties:
      id:
        type: integer
        description: Identifiant
        example: 4
      name:
        type: string
        description: Nom du rôle
        example: Opérateur
  Stade:
    type: object
    description: Stade de production
    properties:
      id:
        type: integer
        description: Identifiant
        example: 4
      name:
        type: string
        description: Nom du stade
        example: Défruitage
      sections:
        type: array
        items:
          $ref: '#/definitions/Section' 
      imgUrl:
        type: string
        description: URL of the image that represents this Stade
        example: http://example.com/bulldozer.png
  Section:
    type: object
    description: Section
    properties:
      id:
        type: integer
        description: Identifiant
        example: 4
      name:
        type: string
        description: Nom de la section
        example: Mécanique
  User:
    type: object
    properties:
      id:
        type: integer
        description: The user ID.
        example: 4
      username:
        type: string
        description: The user name.
        example: rcriton
      firstname:
        type: string
        description: Prénom
        example: Romain
      lastname:
        type: string
        description: Nom de famille
        example: Criton
      email:
        type: string
        description: Adresse e-mail
        example: rcriton@example.com
      role:
        $ref: '#/definitions/Role'
      stade:
        $ref: '#/definitions/Stade'
      section:
        $ref: '#/definitions/Section'
      photo:
        type: string
        description: URL de la photo de l'utilisateur
        example: 'http://www.example.com/photo.jpg'
...